Fixed more warnings.

+++ tests/test.c
@@ -52,7 +52,7 @@ int main(int argc, char *argv[])
   else
     wildcard= NULL;
 
-  srandom(time(NULL));
+  srandom((unsigned int)time(NULL));
 
   for (next= collection; next->name; next++)
   {
@@ -144,9 +144,9 @@ long int timedif(struct timeval a, struct timeval b)
 {
   register int us, s;
 
-  us = a.tv_usec - b.tv_usec;
+  us = (int)(a.tv_usec - b.tv_usec);
   us /= 1000;
-  s = a.tv_sec - b.tv_sec;
+  s = (int)(a.tv_sec - b.tv_sec);
   s *= 1000;
   return s + us;
 }
